gearup/requirements.txt
asgiref==3.8.1
astroid==3.2.2
boto3==1.34.110
botocore==1.34.110
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
dill==0.3.8
Django==5.0.6
django-cors-headers==4.3.1
django-reversion==5.0.12
django4-rest-swagger==1.1.0
djangorestframework==3.15.1
djangorestframework-swagger==0.2.2
GDAL==3.9.0
idna==3.7
install==1.3.5
isort==5.13.2
itypes==1.2.0
Jinja2==3.1.4
jmespath==1.0.1
logilab-astng==0.24.3
logilab-common==2.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==2.0.0rc1
openapi-codec==1.3.2
platformdirs==4.2.2
psycopg==3.1.19
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.32.2
rest-framework-simplejwt==0.0.2
rest_framework_swagger_zl==0.0.4
s3transfer==0.10.1
setuptools==70.0.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.0
tomlkit==0.12.5
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
wheel==0.43.0
